refactor: move header and footer link definitions to a separate links file for better organization
